+++ source/games/exhumed/src/sound.cpp
@@ -678,7 +678,7 @@ void CheckAmbience(int nSector)
     if (!SoundEnabled()) return;
     if (SectSound[nSector] != -1)
     {
-        short nSector2 = SectSoundSect[nSector];
+        int nSector2 = SectSoundSect[nSector];
         walltype* pWall = &wall[sector[nSector2].wallptr];
         if (!soundEngine->IsSourcePlayingSomething(SOURCE_Ambient, &amb, 0))
         {
