+++ source/games/exhumed/src/switch.cpp
@@ -44,7 +44,7 @@ struct Switch
     short nRunPtr;
     int nSector;
     short nRun2;
-    short nWall;
+    int nWall;
     short nRun3;
     uint16_t nKeyMask;
 };
@@ -288,7 +288,7 @@ void AISWStepOn::TouchFloor(RunListEvent* ev)
 
     if (var_14 != sRunChannels[nChannel].c)
     {
-        short nWall = sector[nSector].wallptr;
+        int nWall = sector[nSector].wallptr;
         PlayFXAtXYZ(StaticSound[nSwitchSound], wall[nWall].x, wall[nWall].y, sector[nSector].floorz, nSector);
 
         assert(sRunChannels[nChannel].c < 8);
@@ -450,7 +450,7 @@ void AISWPressSector::Use(RunListEvent* ev)
 
 }
 
-std::pair<int, int> BuildSwPressWall(short nChannel, short nLink, short nWall)
+std::pair<int, int> BuildSwPressWall(short nChannel, short nLink, int nWall)
 {
     if (SwitchCount <= 0 || nLink < 0 || nWall < 0) {
         I_Error("Too many switches!\n");
@@ -482,7 +482,7 @@ void AISWPressWall::Process(RunListEvent* ev)
 
     if (LinkMap[nLink].v[sRunChannels[nChannel].c] >= 0)
     {
-        short nWall = SwitchData[nSwitch].nWall;
+        int nWall = SwitchData[nSwitch].nWall;
         SwitchData[nSwitch].nRun3 = runlist_AddRunRec(wall[nWall].lotag - 1, &RunData[ev->nRun]);
     }
 }
@@ -505,7 +505,7 @@ void AISWPressWall::Use(RunListEvent* ev)
         SwitchData[nSwitch].nRun3 = -1;
     }
 
-    short nWall = SwitchData[nSwitch].nWall;
+    int nWall = SwitchData[nSwitch].nWall;
     int nSector =SwitchData[nSwitch].nSector; // CHECKME - where is this set??
 
     PlayFXAtXYZ(StaticSound[nSwitchSound], wall[nWall].x, wall[nWall].y, 0, nSector, CHANF_LISTENERZ);
